+++ .php-cs-fixer.dist.php
@@ -11,7 +11,9 @@ $config = new PhpCsFixer\Config();
 return $config->setRules([
     '@PER-CS2.0' => true,
     '@PER-CS2.0:risky' => true,
-    'fully_qualified_strict_types' => true,
+    'fully_qualified_strict_types' => [
+        'import_symbols' => true,
+    ],
     'global_namespace_import' => [
         'import_classes' => false,
         'import_constants' => false,