+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * chore: naming, etc
 * test: property path negative test
 * fix: mapping from `stdClass` to object with mandatory constructor arguments
+* feat: variadic setter & constructor arguments support
 
 ## 1.8.0